Add supplement file so I can add my own words
NOTE: This is very fragile since I'm using 'tw' as a key, when it's
actually just an alias for the offial 'zh_Hant' but I want this out
asap. Also had a surprising amount of trouble with the path function...
really should add tests if this project continues in such a way that I
update the core code often.